moved encryption to db-modules
